Split out agent and server python bits
- Move configtools to common directory
- Move pbench-server python modules, classes, etc to pbench/server
- Updated binaries
