+++ pbench/common/exception.py
