+++ pbench/server/indexer.py
@@ -34,8 +34,8 @@ except ImportError:
 # We import the entire pbench module so that mocking time works by changing
 # the _time binding in the pbench module for unit tests via the PbenchConfig
 # constructor's execution.
-import pbench
-from pbench.exception import (
+import pbench.server
+from pbench.common.exception import (
     BadDate,
     ConfigFileError,
     BadMDLogFormat,
@@ -44,8 +44,8 @@ from pbench.exception import (
     MappingFileError,
     TemplateError,
 )
-from pbench.logger import get_pbench_logger
-from pbench.utils import tstos
+from pbench.server.logger import get_pbench_logger
+from pbench.server.utils import tstos
 
 try:
     from .mock import MockElasticsearch
@@ -587,7 +587,7 @@ def es_put_template(es, name=None, body=None):
     retry = True
     retry_count = 0
     backoff = 1
-    beg, end = pbench._time(), None
+    beg, end = pbench.server._time(), None
     # Derive the mapping name from the template name
     mapping_name = "pbench-{}".format(name.split(".")[2])
     try:
@@ -640,7 +640,7 @@ def es_put_template(es, name=None, body=None):
             retry_count += 1
         else:
             retry = False
-    end = pbench._time()
+    end = pbench.server._time()
     return beg, end, retry_count
 
 
@@ -702,7 +702,7 @@ def es_index(es, actions, errorsfp, logger, _dbg=0):
                 # pounding on the ES instance.
                 backoff += 1
 
-    beg, end = pbench._time(), None
+    beg, end = pbench.server._time(), None
     successes = 0
     duplicates = 0
     failures = 0
@@ -811,7 +811,7 @@ def es_index(es, actions, errorsfp, logger, _dbg=0):
                     )
                     actions_retry_deque.append((retry_count + 1, action))
 
-    end = pbench._time()
+    end = pbench.server._time()
 
     assert len(actions_deque) == 0
     assert len(actions_retry_deque) == 0
@@ -4071,7 +4071,7 @@ class IdxContext(object):
         self.name = name
         self._dbg = _dbg
         self.opctx = []
-        self.config = pbench.PbenchConfig(options.cfg_name)
+        self.config = pbench.server.PbenchConfig(options.cfg_name)
         try:
             self.idx_prefix = self.config.get("Indexing", "index_prefix")
         except Exception as e:
@@ -4086,7 +4086,7 @@ class IdxContext(object):
         # We expose the pbench module's internal _time() method here for
         # convenience, allowing us to more easily mock out "time" for unit
         # test environments.
-        self.time = pbench._time
+        self.time = pbench.server._time
         if self.config._unittests:
             import collections
