+++ server/bin/pbench-index
@@ -14,7 +14,7 @@ import tempfile
 from argparse import ArgumentParser
 from configparser import Error as ConfigParserError
 
-from pbench.exception import (
+from pbench.common.exception import (
     BadConfig,
     BadDate,
     BadMDLogFormat,
@@ -24,9 +24,9 @@ from pbench.exception import (
     TemplateError,
     UnsupportedTarballFormat,
 )
-from pbench.indexer import VERSION, IdxContext, es_index, PbenchTarBall
-from pbench.report import Report
-from pbench.utils import quarantine, rename_tb_link, tstos
+from pbench.server.indexer import VERSION, IdxContext, es_index, PbenchTarBall
+from pbench.server.report import Report
+from pbench.server.utils import quarantine, rename_tb_link, tstos
 
 # Internal debugging flag.
 _DEBUG = 0
