+++ server/bin/pbench-verify-backup-tarballs
@@ -46,12 +46,12 @@ import sys
 import tempfile
 from enum import Enum
 
-from pbench import PbenchConfig
-from pbench.exception import BadConfig
-from pbench.logger import get_pbench_logger
-from pbench.utils import md5sum
-from pbench.report import Report
-from pbench.s3backup import Entry, S3Config
+from pbench.server import PbenchConfig
+from pbench.common.exception import BadConfig
+from pbench.server.logger import get_pbench_logger
+from pbench.server.utils import md5sum
+from pbench.server.report import Report
+from pbench.server.s3backup import Entry, S3Config
 
 _NAME_ = "pbench-verify-backup-tarballs"
